Split user helper into actions and reducer.
