+++ assets/helpers/user/userReducer.js
@@ -1,5 +1,10 @@
 // @flow
 
+// ----- Imports ----- //
+
+import type { Action } from './userActions';
+
+
 // ----- Types ----- //
 
 type User = {
@@ -8,11 +13,6 @@ type User = {
   firstName: ?string,
   lastName: ?string,
 };
-
-type Action =
-  | { type: 'SET_FIRST_NAME', name: string }
-  | { type: 'SET_LAST_NAME', name: string }
-  ;
 
 
 // ----- Setup ----- //
@@ -30,18 +30,7 @@ const initialState: User = window.guardian && window.guardian.user ? {
 };
 
 
-// ----- Actions ----- //
-
-export function setFirstName(name: string): Action {
-  return { type: 'SET_FIRST_NAME', name };
-}
-
-export function setLastName(name: string): Action {
-  return { type: 'SET_LAST_NAME', name };
-}
-
-
-// ----- Exports ----- //
+// ----- Reducer ----- //
 
 export default function userReducer(
   state: User = initialState,
